requirements.txt
# Core dependencies
qiskit==0.45.0
qiskit-aer==0.13.0
qiskit-ibmq-provider==0.20.2

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Image processing
Pillow>=9.5.0
scikit-image>=0.20.0
opencv-python>=4.7.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine learning (for clustering)
scikit-learn>=1.2.0

# DNA sequence analysis
biopython>=1.81

# Optimization
cvxpy>=1.3.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Jupyter notebooks
jupyter>=1.0.0
jupyterlab>=3.6.0
ipywidgets>=8.0.0

# Documentation
sphinx>=6.2.0
sphinx-rtd-theme>=1.2.0

# Progress bars
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# Optional: GPU acceleration for Aer
# qiskit-aer-gpu==0.13.0  # Uncomment if you have NVIDIA GPU with CUDA
